Introduce dependency injection:  * seperate dropwizard-jobs-core and dropwizard-jobs-guice modules  * upgraded dropwizard version to 0.7.0-rc2  * Since defaultRegistry has been removed, now using SharedMetricRegistries to shared dropwizard environment metrics registry during bundle initialization.  * updated quartz to 2.2.1  * Introduce Guice DI  * Introduce Spring DI  * Applied mongojack format from @yunspace
