+++ .gitignore
@@ -1,6 +1,6 @@
 .idea
 .classpath
 .project
-dropwizard-jobs.iml
+*.iml
 target/
 .settings
