+++ dropwizard-jobs-core/src/main/java/de/spinscale/dropwizard/jobs/annotations/OnApplicationStart.java
