+++ dropwizard-jobs-core/src/test/java/de/spinscale/dropwizard/jobs/ApplicationStopTestJob.java
