+++ dropwizard-jobs-core/src/test/java/de/spinscale/dropwizard/jobs/OnTestJob.java
