+++ dropwizard-jobs-guice/pom.xml
@@ -1,0 +1,23 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>de.spinscale.dropwizard</groupId>
+        <artifactId>dropwizard-jobs</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>dropwizard-jobs-guice</artifactId>
+    <name>dropwizard-jobs-guice</name>
+    <dependencies>
+        <dependency>
+            <groupId>de.spinscale.dropwizard</groupId>
+            <artifactId>dropwizard-jobs-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.inject.extensions</groupId>
+            <artifactId>guice-servlet</artifactId>
+            <version>3.0</version>
+        </dependency>
+    </dependencies>
+</project>
