+++ dropwizard-jobs-guice/src/test/java/de/spinscale/dropwizard/jobs/Dependency.java
@@ -1,0 +1,5 @@
+package de.spinscale.dropwizard.jobs;
+
+public class Dependency {
+
+}
