+++ dropwizard-jobs-spring/pom.xml
@@ -1,0 +1,33 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>de.spinscale.dropwizard</groupId>
+        <artifactId>dropwizard-jobs</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>dropwizard-jobs-spring</artifactId>
+    <name>dropwizard-jobs-spring</name>
+    <dependencies>
+        <dependency>
+            <groupId>de.spinscale.dropwizard</groupId>
+            <artifactId>dropwizard-jobs-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-core</artifactId>
+            <version>4.0.2.RELEASE</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>4.0.2.RELEASE</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <version>1</version>
+        </dependency>
+    </dependencies>
+</project>
