+++ dropwizard-jobs-spring/src/test/java/de/spinscale/dropwizard/jobs/SpringJobsBundleTest.java
@@ -1,0 +1,36 @@
+package de.spinscale.dropwizard.jobs;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import io.dropwizard.lifecycle.setup.LifecycleEnvironment;
+import io.dropwizard.setup.Environment;
+
+import org.junit.Test;
+import org.mockito.ArgumentCaptor;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+public class SpringJobsBundleTest {
+
+    private final Environment environment = mock(Environment.class);
+    private final LifecycleEnvironment applicationContext = mock(LifecycleEnvironment.class);
+
+    @Test
+    public void assertJobsBundleIsWorking() {
+        ApplicationContext context = new AnnotationConfigApplicationContext(ApplicationStartTestJob.class,
+                ApplicationStopTestJob.class, EveryTestJob.class, OnTestJob.class);
+
+        when(environment.lifecycle()).thenReturn(applicationContext);
+        new SpringJobsBundle(context).run(environment);
+
+        final ArgumentCaptor<JobManager> jobManagerCaptor = ArgumentCaptor.forClass(JobManager.class);
+        verify(applicationContext).manage(jobManagerCaptor.capture());
+
+        JobManager jobManager = jobManagerCaptor.getValue();
+        assertThat(jobManager, is(notNullValue()));
+    }
+}
